Removing questionnaires
The only thing I can possibly come up with is that there’s a bug elsewhere in another package about tracking / something with permissions when it comes to tracking. To deal with this, I’ve just removed the last two questions from each lesson and removed all custom tests that track the student’s progress.